room-agent/pyproject.toml
[project]
name = "room-agent"
version = "0.2.0"
description = "智能机器人代理系统 - 集成语音交互、MCP 工具和 RAG 问答"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.0",
    "langchain-core>=1.0.0",
    "langgraph>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-mcp-adapters>=0.1.0",
    "paho-mqtt>=2.1.0",
]

[project.optional-dependencies]
# BLE Beacon 支持（用于在 Python 端实现 BLE 广播，实际使用 ESP32 硬件）
ble = ["bluepy>=0.2.0"]

# ==================== 告诉 setuptools：这是脚本工程，不是包 ====================
[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
include = []

# ==================== 构建系统 ====================
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

# ==================== 测试 ====================
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

# ==================== 代码格式化 ====================
[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
